feat: audit WIP
